Feature(#3) : 대기방 소켓 컨트롤러 분리 및 Room Create 파라미터 수정
