+++ src/main/java/io/urdego/urdego_game_service/api/room/RoomController.java
@@ -1,44 +1,28 @@
 package io.urdego.urdego_game_service.api.room;
 
-import io.urdego.urdego_game_service.api.room.dto.request.ContentSelectReq;
-import io.urdego.urdego_game_service.api.room.dto.request.PlayerInviteReq;
 import io.urdego.urdego_game_service.api.room.dto.request.RoomCreateReq;
 import io.urdego.urdego_game_service.api.room.dto.response.RoomCreateRes;
-import io.urdego.urdego_game_service.api.room.dto.response.RoomInfoRes;
 import io.urdego.urdego_game_service.domain.room.service.RoomService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
-import org.springframework.stereotype.Controller;
 
-@Controller
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
 @RequiredArgsConstructor
+@RequestMapping("/api/game-service")
 public class RoomController {
 
     private final RoomService roomService;
-    private final SimpMessagingTemplate messagingTemplate;
 
     // 대기방 생성
-    @MessageMapping("/room/create")
-    public void createRoom(RoomCreateReq request) {
+    @PostMapping("/room/create")
+    public ResponseEntity<RoomCreateRes> createRoom(@RequestBody RoomCreateReq request) {
         RoomCreateRes response = roomService.createRoom(request);
-        messagingTemplate.convertAndSend("/game-service/sub" + response.roomId(), response);
+        return ResponseEntity.ok().body(response);
     }
-
-    // 친구 초대
-    @MessageMapping("/room/invite")
-    public void invitePlayer(PlayerInviteReq request) {
-        RoomInfoRes response = roomService.joinRoom(request.roomId(), request.userId());
-        messagingTemplate.convertAndSend("/game-service/sub" + request.roomId(), response);
-    }
-
-    // 컨텐츠 선택 (최대 5개, 0개일 경우 자체 컨텐츠 제공)
-    @MessageMapping("/room/select-content")
-    public void selectContent(ContentSelectReq request) {
-        roomService.registerContents(request);
-    }
-
-    // 준비 상태
-
 
 }
