+++ src/main/java/io/urdego/urdego_game_service/api/room/dto/request/RoomCreateReq.java
@@ -3,9 +3,9 @@ package io.urdego.urdego_game_service.api.room.dto.request;
 import java.util.List;
 
 public record RoomCreateReq(
+        String userId,
         int maxPlayers,
         int totalRounds,
-        int timer,
-        List<String> currentPlayers
+        int timer
 ) {
 }
